Source > CleanUp.

+++ MODSRC/vazkii/recubed/common/core/handler/GeneralEventHandler.java
@@ -1,11 +1,11 @@
 /**
  * This class was created by <Vazkii>. It's distributed as
  * part of the ReCubed Mod.
- * 
+ *
  * ReCubed is Open Source and distributed under a
  * Creative Commons Attribution-NonCommercial-ShareAlike 3.0 License
  * (http://creativecommons.org/licenses/by-nc-sa/3.0/deed.en_GB)
- * 
+ *
  * File Created @ [Dec 13, 2013, 3:23:32 PM (GMT)]
  */
 package vazkii.recubed.common.core.handler;
@@ -30,7 +30,6 @@ import net.minecraftforge.event.ForgeSubscribe;
 import net.minecraftforge.event.ServerChatEvent;
 import net.minecraftforge.event.entity.item.ItemTossEvent;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
-import net.minecraftforge.event.entity.living.LivingEvent.LivingJumpEvent;
 import net.minecraftforge.event.entity.living.LivingHurtEvent;
 import net.minecraftforge.event.entity.living.LivingSetAttackTargetEvent;
 import net.minecraftforge.event.entity.player.ArrowLooseEvent;
@@ -49,13 +48,13 @@ public final class GeneralEventHandler {
 	// ARROWS SHOT
 	@ForgeSubscribe(priority = EventPriority.LOWEST)
 	public void onEntityTakeDamage(ArrowLooseEvent event) {
-        float f = (float) event.charge / 20.0F;
+        float f = event.charge / 20.0F;
         f = (f * f + f * 2.0F) / 3.0F;
-        
+
         if(ReCubedAPI.validatePlayer(event.entityPlayer))
 			ReCubedAPI.addValueToCategory(LibCategories.ARROWS_SHOT, event.entityPlayer.username, f >= 1F ? "recubed.misc.critical_shot" : "recubed.misc.shot", 1);
 	}
-	
+
 	// BLOCKS BROKEN
 	@ForgeSubscribe(priority = EventPriority.LOWEST)
 	public void onBlockBroken(BlockEvent.BreakEvent event) {
@@ -63,7 +62,7 @@ public final class GeneralEventHandler {
 			ReCubedAPI.addValueToCategory(LibCategories.BLOCKS_BROKEN, event.getPlayer().username, Item.itemsList[event.block.blockID].getUnlocalizedName(new ItemStack(event.block.blockID, 1, event.blockMetadata)) + ".name", 1);
 	}
 
-	
+
 	// DAMAGE DEALT
 	@ForgeSubscribe(priority = EventPriority.LOWEST)
 	public void onEntityTakeDamage(LivingHurtEvent event) {
@@ -91,14 +90,14 @@ public final class GeneralEventHandler {
 				ReCubedAPI.addValueToCategory(LibCategories.DAMAGE_TAKEN, player.username, name, (int) event.ammount);
 		}
 	}
-	
+
 	// ITEMS BROKEN
 	@ForgeSubscribe(priority = EventPriority.LOWEST)
 	public void onItemBroken(PlayerDestroyItemEvent event) {
 		if(ReCubedAPI.validatePlayer(event.entityPlayer))
 			ReCubedAPI.addValueToCategory(LibCategories.ITEMS_BROKEN, event.entityPlayer.username, event.original.getUnlocalizedName() + ".name", 1);
 	}
-	
+
 	// ITEMS DROPPED
 	@ForgeSubscribe(priority = EventPriority.LOWEST)
 	public void onPlayerTossItem(ItemTossEvent event) {
@@ -127,7 +126,7 @@ public final class GeneralEventHandler {
 	public void onMessageReceived(CommandEvent event) {
 		if(event.sender instanceof EntityPlayer && ReCubedAPI.validatePlayer((EntityPlayer) event.sender)) {
 			ReCubedAPI.addValueToCategory(LibCategories.MESSAGES_SENT, event.sender.getCommandSenderName(), "recubed.misc.command", 1);
-			
+
 			if(event.command instanceof CommandGive) {
 				int i = CommandBase.parseIntWithMin(event.sender, event.parameters[1], 1);
 	            int j = 1;
@@ -135,7 +134,7 @@ public final class GeneralEventHandler {
 
 	            if (Item.itemsList[i] == null)
 	                return;
-	           
+
                 if (event.parameters.length >= 3)
                     j = CommandBase.parseIntBounded(event.sender, event.parameters[2], 1, 64);
 
@@ -143,12 +142,12 @@ public final class GeneralEventHandler {
                     k = CommandBase.parseInt(event.sender, event.parameters[3]);
 
                 ItemStack stack = new ItemStack(i, j, k);
-                
+
     			ReCubedAPI.addValueToCategory(LibCategories.ITEMS_SPAWNED, event.sender.getCommandSenderName(), stack.getUnlocalizedName() + ".name", 1);
 			}
 		}
 	}
-	
+
 	// MOBS AGGROED
 	@ForgeSubscribe(priority = EventPriority.LOWEST)
 	public void onMobGetTarget(LivingSetAttackTargetEvent event) {
@@ -170,7 +169,7 @@ public final class GeneralEventHandler {
 				ReCubedAPI.addValueToCategory(event.entity instanceof IBossDisplayData ? LibCategories.BOSS_KILLS : LibCategories.MOBS_KILLED, player.username, name, 1);
 		}
 	}
-	
+
 
 	// TIMES DIED + PLAYER KILLS
 	@ForgeSubscribe(priority = EventPriority.LOWEST)
@@ -190,20 +189,20 @@ public final class GeneralEventHandler {
 				ReCubedAPI.addValueToCategory(LibCategories.TIMES_DIED, player.username, name, 1);
 		}
 	}
-	
+
 	// SNOWBALLS THROWN + ENDER PEARLS THROWN
 	@ForgeSubscribe(priority = EventPriority.LOWEST)
 	public void onPlayerInteract(PlayerInteractEvent event) {
 		if(ReCubedAPI.validatePlayer(event.entityPlayer)) {
 			ItemStack stack = event.entityPlayer.getCurrentEquippedItem();
-			
+
 			if(stack != null) {
 				if(stack.itemID == Item.snowball.itemID)
 					ReCubedAPI.addValueToCategory(LibCategories.SNOWBALLS_THROWN, event.entityPlayer.username, "item.snowball.name", 1);
-				
+
 				if(!event.entityPlayer.capabilities.isCreativeMode && stack.itemID == Item.enderPearl.itemID)
 					ReCubedAPI.addValueToCategory(LibCategories.ENDER_PEARLS_THROWN, event.entityPlayer.username, "item.enderPearl.name", 1);
-			
+
 				if(stack.getItem() instanceof ItemRecord && event.action == Action.RIGHT_CLICK_BLOCK && event.entityPlayer.worldObj.getBlockId(event.x, event.y, event.z) == Block.jukebox.blockID)
 					ReCubedAPI.addValueToCategory(LibCategories.DISCS_PLAYED, event.entityPlayer.username, ((ItemRecord) stack.getItem()).recordName, 1);
 			}
@@ -216,7 +215,7 @@ public final class GeneralEventHandler {
 		if(!ReCubedAPI.validatePlayer(event.entityPlayer))
 			return;
 
-		
+
 		EnumStatus status = event.result;
 		if(status == null) {
 			findStatus : {
@@ -225,29 +224,29 @@ public final class GeneralEventHandler {
 						status = EnumStatus.OTHER_PROBLEM;
 						break findStatus;
 					}
-	
+
 					if (!event.entityPlayer.worldObj.provider.isSurfaceWorld()) {
 						status = EnumStatus.NOT_POSSIBLE_HERE;
 						break findStatus;
 					}
-	
+
 					if (event.entityPlayer.worldObj.isDaytime()) {
 						status = EnumStatus.NOT_POSSIBLE_NOW;
 						break findStatus;
 					}
-	
-					if (Math.abs(event.entityPlayer.posX - (double) event.x) > 3D || Math.abs(event.entityPlayer.posY - (double) event.y) > 3D || Math.abs(event.entityPlayer.posZ - (double) event.z) > 3D) {
+
+					if (Math.abs(event.entityPlayer.posX - event.x) > 3D || Math.abs(event.entityPlayer.posY - event.y) > 3D || Math.abs(event.entityPlayer.posZ - event.z) > 3D) {
 						status = EnumStatus.TOO_FAR_AWAY;
 						break findStatus;
 					}
-	
+
 					double d0 = 8.0D;
 					double d1 = 5.0D;
-					List list = event.entityPlayer.worldObj.getEntitiesWithinAABB(EntityMob.class, AxisAlignedBB.getAABBPool().getAABB((double) event.x - d0, (double) event.y - d1, (double) event.z - d0, (double) event.x + d0, (double) event.y + d1, (double) event.z + d0));
-	
+					List list = event.entityPlayer.worldObj.getEntitiesWithinAABB(EntityMob.class, AxisAlignedBB.getAABBPool().getAABB(event.x - d0, event.y - d1, event.z - d0, event.x + d0, event.y + d1, event.z + d0));
+
 					if (!list.isEmpty()) {
 						status = EnumStatus.NOT_SAFE;
-						break findStatus; 
+						break findStatus;
 					}
 					status = EnumStatus.OK;
 				}
